+++ src/main/java/com/elmakers/mine/bukkit/plugins/spells/builtin/InvincibleSpell.java
@@ -6,19 +6,48 @@ import com.elmakers.mine.bukkit.plugins.spells.Spell;
 
 public class InvincibleSpell extends Spell 
 {
+    public InvincibleSpell()
+    {
+        addVariant("ironskin", Material.IRON_CHESTPLATE, getCategory(), "Protect you from damage", "99");
+        addVariant("leatherskin", Material.LEATHER_CHESTPLATE, getCategory(), "Protect you from some damage", "50");
+    }
+    
 	@Override
 	public boolean onCast(String[] parameters) 
 	{
-		boolean invincible = !spells.isInvincible(player);
-		spells.setInvincible(player, invincible);
-		if (invincible)
-		{
-			castMessage(player, "You feel invincible!");
-		}
-		else
-		{
-			castMessage(player, "You feel ... normal.");
-		}
+	    int amount = 100;
+        if (parameters.length > 0)
+        {
+            try
+            {
+                amount = Integer.parseInt(parameters[0]);
+            }
+            catch (NumberFormatException ex)
+            {
+                amount = 100;
+            }
+        }
+        
+        Float currentAmount = spells.invincibleAmount(player);
+        if (currentAmount != null)
+        {
+            sendMessage(player, "You feel ... normal.");
+            spells.setInvincible(player, 0);
+        }
+        else
+        {
+            spells.setInvincible(player, (float)amount / 100);
+            
+            if (amount >= 100)
+            {
+                sendMessage(player, "You feel invincible!");
+            }
+            else
+            {
+                sendMessage(player, "You feel strong!");
+            }
+        }
+       
 		return true;
 	}